+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog for Meteor Mogul meteor-accounts-vue-unstyled
 
+## [0.0.5] - 2018-01-10
+
+### Changed
+
+* Bumped semver to force use of updated package
+
+## [0.0.4] - 2018-01-10
+
+### Added
+
+* Export LoginButtons Vue component
+
+### Changed
+
+* Change name of package to `accounts-vue-unstyled`
+* `master` branch now calling package `accounts-vue-unstyled`
+* Use `meteormogul:vue-dist@2.5.15` which exports Vue symbol
+
 ## 0.0.3 - (2018-01-03)
 
 ### Changed